+++ MatchEntities/Models/BombExplosion.cs
@@ -11,7 +11,7 @@ namespace MatchEntities
         IIngameEvent
     {
         public long MatchId { get; set; }
-        public short Round { get; set; }
+        public byte Round { get; set; }
         public int Time { get; set; }
 
         [JsonIgnore]
